requirements.txt
google-generativeai
matplotlib
seaborn
pandas